V1.5.2/2
[CURRENT]
Tested snapping script for pipe puzzle, works but needs more work.
Started scripts for detection of puzzles being done.

[NOTES]
Need to make script compare the currently snapped-to point's bool whether it's false or not; if it's true, it'll emit a signal of 1, which will be picked up by the game controller.